+++ CHANGELOG.md
@@ -3,6 +3,8 @@
  - #455 - fix the `altorouter` link in create new page
  - fix `altorouter` links in `impressum`
  - #458 - fix delete page bug
+ - #459 - do not load the links in CMS edit mode
+ - #453 - set default gender to `admin`
 
 # v6.10.0
 ### New features